Add fit_run method to dag
